pyproject.toml
[tool.poetry]
name = "simple-eval"
version = "0.1.0"
description = "Lightweight chatbot evaluation platform"
authors = ["User"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
aiosqlite = "^0.19.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve = "main:main"